Don't copy the TI generic nvs bin as we will calibrate anyway
This will ensure the proper MAC is written after dirty flash

+++ prebuilt/bin/wifical.sh
@@ -33,7 +33,10 @@ fi
 mount -o remount rw /system
 
 # Remove old NVS file
+if [ -e $TARGET_NVS_FILE ];
+then
 rm /system/etc/firmware/ti-connectivity/wl1271-nvs.bin
+fi
 
 # Actual calibration...
 # calibrator plt autocalibrate <dev> <module path> <ini file1> <nvs file> <mac addr>
